JavaScript Use Strict
The "use strict" directive is new in JavaScript 1.8.5 (ECMAScript version 5).
It is not a statement, but a literal expression, ignored by earlier versions of JavaScript.
The purpose of "use strict" is to indicate that the code should be executed in "strict mode".
With strict mode, you can not, for example, use undeclared variables.
More info at https://www.w3schools.com/js/js_strict.asp

Signed-off-by: Phạm Tiến Thành <tienthanh.dqc@gmail.com>